requirements.txt
openai>=1.0.0
anthropic>=0.19.1
pinecone-client>=3.0.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0
click>=8.0.0
pytest>=7.0.0
urllib3>=2.0.0
